Move reco and tracking factories into factories dir
